requirements.txt
tensorflow-cpu
opencv-python
pillow
uvicorn
gunicorn
fastapi
flask
librosa
python-multipart